+++ src/main/java/waffle/nbayes/NBayesClassifierResult.java
@@ -8,9 +8,9 @@ import waffle.core.ClassifierResult;
 public class NBayesClassifierResult implements ClassifierResult {
 
     private String matchedCategory;
-    private Map<String, Float> scoresByCategories;
+    private Map<String, Double> scoresByCategories;
     
-    public NBayesClassifierResult(Map<String, Float> scoresByCategories) {
+    public NBayesClassifierResult(Map<String, Double> scoresByCategories) {
         this.scoresByCategories = scoresByCategories;
     }
     
@@ -27,14 +27,14 @@ public class NBayesClassifierResult implements ClassifierResult {
         return scoresByCategories.keySet();
     }
     
-    public float getScoreForCategory(String category) {
+    public double getScoreForCategory(String category) {
         return scoresByCategories.get(category);
     }
     
     private String getCategoryWithBestScore() {
-        Map.Entry<String, Float> bestScoreEntry = null;
+        Map.Entry<String, Double> bestScoreEntry = null;
         
-        for (Map.Entry<String, Float> scoreEntry : scoresByCategories.entrySet()) {
+        for (Map.Entry<String, Double> scoreEntry : scoresByCategories.entrySet()) {
             
             if (bestScoreEntry == null || bestScoreEntry.getValue() < scoreEntry.getValue() ) {
                 bestScoreEntry = scoreEntry;
